Refactor compute package structure
