+++ src/main/java/com/iexec/worker/compute/DockerExecutionConfig.java
@@ -1,4 +1,4 @@
-package com.iexec.worker.docker;
+package com.iexec.worker.compute;
 
 import com.iexec.common.utils.ArgsUtils;
 import lombok.AllArgsConstructor;
