+++ src/test/java/com/iexec/worker/executor/TaskManagerServiceTests.java
@@ -18,10 +18,10 @@ import com.iexec.common.task.TaskDescription;
 import com.iexec.worker.chain.ContributionService;
 import com.iexec.worker.chain.IexecHubService;
 import com.iexec.worker.chain.RevealService;
+import com.iexec.worker.compute.ComputationService;
+import com.iexec.worker.compute.CustomDockerClient;
 import com.iexec.worker.config.WorkerConfigurationService;
 import com.iexec.worker.dataset.DataService;
-import com.iexec.worker.docker.ComputationService;
-import com.iexec.worker.docker.CustomDockerClient;
 import com.iexec.worker.feign.CustomCoreFeignClient;
 import com.iexec.worker.result.ResultService;
 import com.iexec.worker.sms.SmsService;
